cc1plus: all warnings being treated as errors fix
1) error: missing initializer for member OicSecKey::len
2) ISO C++ forbids converting a string constant to 'char*'
3) _POSIX_C_SOURCE redefined

+++ resource/csdk/security/provisioning/sample/autoprovisioningclient.c
@@ -173,7 +173,7 @@ static void provisionCredCB(void* ctx, size_t nOfRes, OCProvisionResult_t* arr, 
 }
 
 /* Function of type OCProvisionResultCB from \resource\csdk\security\provisioning\include\pmtypes.h */
-void provisionTrustChainCB(void* ctx, int nOfRes, OCProvisionResult_t *arr, bool hasError)
+void provisionTrustChainCB(void* ctx, size_t nOfRes, OCProvisionResult_t *arr, bool hasError)
 {
     if(!hasError)
     {
@@ -571,9 +571,9 @@ static int doGetRequest(const char* uri, int dev_num)
         return -1;
     }
 
-    /* 
-     * Make sure the OCDevAddr passed to OCDoRequest contains the remote peer's UUID, 
-     * so that PSK lookup succeeds. 
+    /*
+     * Make sure the OCDevAddr passed to OCDoRequest contains the remote peer's UUID,
+     * so that PSK lookup succeeds.
      */
     if (!OCConvertUuidToString(device->doxm->deviceID.id, device->endpoint.remoteId))
     {
